Issue-31414: Unify flashinfer utils at different paths
